# Curl sb (single-batched)