--- a/curl/classic/curl_suite_test.go
+++ b/curl/classic/curl_suite_test.go
@@ -1,4 +1,4 @@
-package curl_test
+package classic_test
 
 import (
 	"testing"