--- a/curl/classic/transform_decl.go
+++ b/curl/classic/transform_decl.go
@@ -1,7 +1,9 @@
 // +build !gccgo,!appengine
 // +build amd64
 
-package curl
+package classic
+
+import "github.com/iotaledger/iota.go/curl"
 
 //go:noescape
-func transform(dst, src *[StateSize]int8, rounds uint)
+func transform(dst, src *[curl.StateSize]int8, rounds uint)