… curl/classic/curl_test.go
… curl/classic/transform_amd64.s
… curl/curl_internal_test.go
… curl/curl_suite_test.go
… curl/curl_test.go