# Description of change

Closes issue #198 .
Reduce energy consumption of Curl single hasher, by applying the curl function in a batched manner.
If you know curl hashing, it's quite obvious for the first round: you can rotate the state by 364, then use opposite trits as inputs for the curl function. The resulting state will have resorted indices, but their offsets are consistent throughout the new state. 
Rotate again to have index 364 being in opposition to index 0, then again 'curl' them. Repeat, until rounds are done.
Finally, sort the state.
Similar to bct, it encodes one trit as two bits on different uint64 arrays,
n (Negative) and p (Positive): 0 -> n=0, p=0; 1 -> n=0, p=1; -1 -> n=1, p=0.
Top level arrays store 243 trits each, 2nd level arrays  get filled up 3*64+51.
This way, when doing 81 rounds, resorting can be done by and-mask operations.

## Type of change

- Enhancement (a non-breaking change which adds functionality)

## How the change has been tested

Besides additional ginkgo tests, I'm running a HORNET node with the change successfully so far.

## Change checklist

- [x] My code follows the contribution guidelines for this project
- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas
- [x] I have added tests using ginkgo that prove my fix is effective or that my feature works
- [x] New and existing unit tests pass locally with my changes