@Wollac 2020-11-10
This idea is really impressive!

I've re-implemented your proposal here https://github.com/Wollac/iota.go/tree/curl-sb 
It tries to address the most relevant issues I see with this PR: I believe, it improves code readability by being more "idiomatic go". This approach is really tailored for 81-round Curl, so it removes the deprecated Curl-P-27 completely. It does not rely on C-code to improve maintainability and portability.

By doing so, https://github.com/Wollac/iota.go/tree/curl-sb still achieves a 4x-5x speedup on amd64 over current `master`: 
```
name                old time/op  new time/op  delta
CurlTransaction     1.53ms ± 3%  0.32ms ± 2%  -79.27%  (p=0.000 n=12+11)
CurlHash            48.2µs ± 3%  11.9µs ± 5%  -75.27%  (p=0.000 n=12+12)
```

Of course, AVX2 Cgo code in #199 are still a bit faster. (I've measured 10%-15% on my machine.) But a cleaner native go implementation might justify this small drawback.