@c-mnd 2020-11-10
Thank You. I knew, the algorithm is ok, but my implementation was of course lousy.
I'll read your rewrite and learn!