@c-mnd 2020-11-11
Yeah, this way it's a lot cleaner.
Compiler hints like these would have never come to my mind, great.
What I don't get: calling it a _left_ rotation, while the 'leading' shift is `shrInto`. Is it intentionally (looking at it as a higher level array context)?