@Wollac 2020-11-11
Yeah, what I mean is that when calling `rotateState` with `1`, it will move the trit at index i to index i-1. So, the state elements are rotated to the left, which is exactly what you want when comparing it to the array based `transform`.
However, you are absolutely right, on a bit level it is doing a `ROR`.