@c-mnd 2020-11-11
In an earlier implementation I called it downshift (`shd`) when the movement was towards lower indices, upshift (`shu`) for the opposite. This way it should work for both LTR (tits, trytes, enumerations) and conventional RTL numbers, lowlevel as well as higher.
Still, that introduced new terms and the necessity for imaginary translation even grew.
Finally, I just dropped naming the rotations direction, as these are not hard-to-understand code areas.

So the information you might want to provide in the comments is that there's _motion towards lower indices_.