@c-mnd 2020-11-11
When possible and semantics don't suffer, operations on state members should be done in the order of their definition, for better reading and maintaining.
Hopefully I caught them all: https://github.com/c-mnd/iota.go/commit/42363bf8b26917942bfe32a24540473c2317fc59.